there seems to be no non-kludge way to change the app's language at run time
I wanted to create a base class for all of my activities that would allow the user to select Japanese or English at run time, and have it re-load the activity with the new language setting.

I'm just going to leave it be for now. The language can be set to Japanese via the android settings menu